+++ lib/amorail/entity/persistence.rb
